requirements/dev.txt
-r common.txt
gunicorn==0.17.2
psycopg2==2.4.5
Flask-DebugToolbar==0.9.0
